requirements.txt
numpy==1.18.1
pandas==1.0.0
tqdm==4.45.0
matplotlib==3.2.1
pyyaml==5.1
